Replace original proposals API with working beautiful version and clean up test files
